+++ CHANGELOG.md
@@ -2,6 +2,13 @@ Please refer first to [UPGRADE.md](UPGRADE.md) for the most important items that
 
 The current file describes significant changes in the code that may affect the upgrade of your customizations.
 
+## 4.2.0-alpha.3
+
+### Changed
+
+#### UserBundle
+* The name for `/api/authstatuses` REST API resource was changed to `/api/userauthstatuses`.
+
 ## 4.2.0-alpha.2 (2020-05-29)
 [Show detailed list of changes](incompatibilities-4-2-alpha-2.md)